Test including JSON Schema component in packaging tests
