+++ config.cmake.in
@@ -2,5 +2,14 @@
 
 include("${CMAKE_CURRENT_LIST_DIR}/@JSONTOOLKIT_BACKEND@.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_json.cmake")
+
+# GCC does not allow the use of std::promise, std::future
+# without compiling with pthreads support.
+if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
+  set(CMAKE_THREAD_PREFER_PTHREAD TRUE)
+  set(THREADS_PREFER_PTHREAD_FLAG TRUE)
+  find_package(Threads REQUIRED)
+endif()
+
 include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_jsonschema.cmake")
 check_required_components("@PROJECT_NAME@")
