+++ test/packaging/find_package/CMakeLists.txt
@@ -6,3 +6,4 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 find_package(sourcemeta_jsontoolkit REQUIRED)
 add_executable(jsontoolkit_hello hello.cc)
 target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::json)
+target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::jsonschema)
