+++ test/packaging/find_package/hello.cc
@@ -1,4 +1,5 @@
 #include <sourcemeta/jsontoolkit/json.h>
+#include <sourcemeta/jsontoolkit/jsonschema.h>
 
 #include <cstdlib>  // EXIT_SUCCESS
 #include <iostream> // std::cout
